+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Nightly
+- Qwen3-TTS: add audio output, `extract_tts`, `mmproj_use_gpu`, `mmproj_flash_attn`, `mmproj_batch_max_tokens`, `language` config
+- Add `answer_delimiter` config
 - Add `streaming_mode`, refactor subprocess mode - the process is now interruptible.
 - Add speculative decoding
 - Add dynamic image, audio, video input